DateTimes should be Kind.Unspecified, for #2
These DateTimes obviously correspond to a particular Western US local time (this is the City of Santa Monica, CA after all). The problem comes when serializing to JSON in ISO8061 format; specifying as Kind.Local means that depending upon when the data is serialized, the output could be UTC offset -7 or -8 (because of Daylight Savings Time). Specifying as Kind.Utc is also not correct as that is *always* ahead of Santa Monica, CA local time.